Enhance section layout and style
