+++ src/containers/Main/Main.js
@@ -5,7 +5,6 @@ import Login from '../../pages/Login'
 import Signup from '../../pages/Signup'
 import SubmitPost from '../../pages/SubmitPost'
 import styled from 'styled-components'
-import Container from '../../components/Container'
 
 const StyledMain = styled.main`
   flex: 1;
@@ -15,14 +14,12 @@ const StyledMain = styled.main`
 
 const Main = () => (
   <StyledMain>
-    <Container>
-      <Switch>
-        <Route exact path="/" component={Home} />
-        <Route path="/login" component={Login} />
-        <Route path="/signup" component={Signup} />
-        <Route path="/submit" component={SubmitPost} />
-      </Switch>
-    </Container>
+    <Switch>
+      <Route exact path="/" component={Home} />
+      <Route path="/login" component={Login} />
+      <Route path="/signup" component={Signup} />
+      <Route path="/submit" component={SubmitPost} />
+    </Switch>
   </StyledMain>
 )
