+++ src/pages/Home/Home.js
@@ -1,11 +1,15 @@
 import React from 'react'
+import Section from '../../components/Section'
 import PostList from '../../containers/PostList'
+import Container from '../../components/Container'
 
 const Home = () => {
   return (
-    <div>
-      <PostList />
-    </div>
+    <Container>
+      <Section>
+        <PostList />
+      </Section>
+    </Container>
   )
 }
